+++ CHANGELOG.md
@@ -17,7 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `GAPMonitorCallback` for monitoring gradients, activations, and parameters (GAP).
 - Added `PeriNormTransformerBlock`.
 - Added official OLMo3-7B pretraining scripts and data mix.
+- Added official OLMo3-7B and -32B pretraining scripts and data mix.
+- Added official OLMo3-7B and -32B midtraining script and data mix.
+- Added official OLMo3-7B and -32B long-context script and data mix.
 - Added a `NoOpOptimizer` that does nothing, uses no memory, and can be used for debugging.
+- Added official config for Olmo 3 32B.
 
 ### Fixed
 
@@ -26,6 +30,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - No longer overrides `random`'s RNG seed when building `SourceMixtureDatasetConfig`.
 - Fix handling URLs in `olmo_core.nn.hf.checkpoint.save_hf_model` and in `examples/huggingface`.
 - Fix potential NaN loss that can occur when using instance masking.
+- Stability improvements developed while training Olmo3 32B.
+
+### Changed
+
+- Removed unused field in `YaRNRoPEScalingConfig`.
 
 ## [v2.3.0](https://github.com/allenai/OLMo-core/releases/tag/v2.3.0) - 2025-10-17